fsodeviced: kernel26_rtc: be more cooperative with /dev/rtc0
This module will now only hold /dev/rtc0 while it is a) doing any operation
(which is very short), or b) waiting for a pending alarm (in which case
we do not want someone else to remove or change it behind our back anyways).

This should improve FSO #492 to a certain extent. Please open another ticket,
if you are missing something in the RTC DBus API.