fsodeviced: alsa_audio: use inotify to watch for scenario changes
Reload the scenario cache (for the modified scenario only) whenever
a change is being reported. This also works, if the changed scenario
is the current scenario. If the scenario being modified is not the
current scenario, be aware that audible artefacts can occur during
reloading.
NOTE: We require libfsoframework 0.2.4 and Vala 0.7.8.2 now