libgsm0710mux: introduce send_pause_threshold configuration value for unstable modems
This configuration value allows to specify a minimum pause time interval
that has to be guaranteed when sending consecutive commands. Some modems tend to
hang up, if you send commands too fast. Yes, this is what the channel flow control
out-of-band data should be used, but alas, it seems we can't rely on that for all
modems. So we play safe. Note that this value is initially set to 0; change it like:
[libgsm0710mux]
device_sendpause_threshold = 100