fsoinitd: add AdditionalBootProcessAction
Sometimes the distribution have to launch several additional processes while booting. As
we want the boot process as short and small as possible the AdditionBootProcessAction
adds the possibility to start additional processes right after the main boot process. It
is launched in the background so the user see the UI comming up while several other
process are launched.