fsogsmd: modem_qualcomm_palm: implement most mediators for sim/phonebook access
This implements mostly all necessary mediators for the sim/phonebook access. There are
some untested parts which will be tested and verified later. Also there is a new method
named 'waitForUnsolicitedResponse' which lets your current method wait for recieving a
specified unsolicitied response. For some commands like the the SimWriteEntry the result
of the command is not supplied by the solicited response. It is part of the deffered send
unsolicited response PhonebookModified which will recieve a bit later than the solicited
response. There are even some more parts in the msmcomm protocol which has the same
behaviour.