fsodeviced: fix(in configuration) the remote silence when answering a call
The alsa routing resulted in a problematic issue:
*someone calls you
*fso routes the ringing to the external speaker:
  ( The external speaker is usually used for hand-free calling
  or for playing some music loud, it's not used for normal
  phone conversations )
*the phone rings
*you press the answer button in SHR's phone GUI

Then:
You can't ear the remote person talking.
(leviathan said that the remote person could ear you tough)

Removing this routing fix that issue,but the ringing is done
trough the current audio path, in other words:
if you're on headphones, it'll ring on the headphones.
(which can be desirable or not)