fsodeviced: accelerometer: stateless orientation
When flat we still reports the previous (when it was held) screen orientation.
The orientation changed signal now sends the same info as the get_orientation
method, i.e. the full orientations state.  This eliminates the need for the
client to keep track of this state information.

These messages are (also/still) compliant with the current dbus specification.

Signed-off-by: Peter van de Werken <pwerken-fso@a-eskwadraat.nl>