libfsoresource: implement resource registration after registration was lost
In a few cases it can happen that the fsousage daemon dies. When it comes back again all
registrations are lost. Especially when using shadow resources this will let the usage
daemon wait for ever for the resource to register again. The AbstractDBusResource will
now watch if the usage bus disappears and then register the resource again when it comes
back.

In a stable system this case should never happen but when it happens it is better when the
system can solve this simple problem on its own.