fsogsmd: modem_samsung: explicitly cast sim status byte to uint8 to let the switch succeed
Without casting the status before taking it into account within the switch as a uint8 the
switch will fail. While debuging this I saw that modem.status changed to 0xffffff82
somewhere. Thats some indicator for the switch failing when comparing 0xffffff82 and 0x82
but I don't know why the leading bytes are changing to 0xff ...

Until we found out why this happens we have here an explicit cast of the status byte to an
uint8.

Signed-off-by: Simon Busch <morphis@gravedo.de>