libfsosystem: add simple abstraction class for android wakelocks
In general no one really needs this and wants to use it but there are some cases where we
should to integrate the best way with a android based kernel. If this class is used on
systems without wakelock support using it will have no impact on the system suspend
handling.

Signed-off-by: Simon Busch <morphis@gravedo.de>