fsogsmd: modem_option_gtm601: rather than using a combined transport create another channel
If we're treating the additional line on the GTA04 as the same as the primary modem port
we have troubles in our AT parser with unsolicited response comming in between when a AT
command is still pending. To avoid this we have to create an additional channel for the
second line.