fsogsmd: allow for fsonetworkd-less configuration by integration with ppp scripts
This introduces an additional [fsogsm] parameter which is supposed to
be set like this (with ip-up from the ppp package):
pdp_helper = /etc/ppp/ip-up

This allows for a clean integration with other services installed on a
typical system, e.g. resolvconf, ntp etc.

To actually assign an address to the interface one needs an additional
script, an example is provided in conf/ppp/ip-up.d/0000configureif .

This is tested on vanilla Debian unstable system with resolvconf on
gta04.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>