fsogsmd: ti_calypso: do not send @POFF ever
Current resource disabling code depends heavily on the modem being
alive. Do not send @POFF to Calypso to be able to properly release and
then request again the GSM resource.

This is likely to leave Calypso up and running on GTA01 even if the
user doesn't need it, increasing power consumption.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>