fsogsmd: allow to disable URC configuration on specific channels, use for Calypso
Commit fa9cdf0455a5a6ca36b4b30f1c715e4cb318b047 introduced automatic
URC config for +CREG notifications on the main channel.

However, modem_ti_calypso uses a dedicated channel for all URCs and
takes special care to silence them on that channel prior to
suspending. Even though the suspend sequence does disable it on
channel 3 it still wakes up the SoC from suspend coming on channel 2
(main).

This patch allows to add a per-channel flag for individual modems to
specify which channels are suitable for URC (true by default for all
channels) and reverts the previous attempt of fixing this for calypso.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>